pyproject.toml
[project]
name = "novel-writer"
version = "0.1.0"
description = "A multi-agent system for writing coherent long-form novels"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-chroma>=0.2.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "typer>=0.12.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[project.scripts]
novel-writer = "novel_writer.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/novel_writer"]
